refactor(resources): refactor resource loader to remove code duplication
